// src/app/interface.ts
// Copyright (c) 2025 Perpetuator LLC
export interface SiteStatistics {
  totalUsers: number;
  activeUsers30Days: number;
  totalTeams: number;
  totalPodcasts: number;
  enabledPodcasts: number;
  totalEpisodes: number;
  liveEpisodes: number;
  totalNewsArticles: number;
  totalJobs: number;
  pendingJobs: number;
  runningJobs: number;
  completedJobs: number;
  failedJobs: number;
  totalCompanies: number;
  totalDcfAnalyses: number;
  totalStockPrices: number;
  totalVoices: number;
  enabledVoices: number;
  totalAudioMinutesGenerated: number;
  totalAudioMinutesPublished: number;
  recentEpisodes: number;
  recentNewsProcessed: number;
  recentDcfAnalyses: number;
  recentJobs: number;
  recentPendingJobs: number;
  recentRunningJobs: number;
  recentCompletedJobs: number;
  recentFailedJobs: number;
}